refactor: data generation to include only projects and organisations
